make plex less confusing
